+++ src/Provider/MiddlewarePipelineServiceProvider.php
@@ -7,13 +7,13 @@ namespace Maduser\Argon\Middleware\Provider;
 use Maduser\Argon\Container\AbstractServiceProvider;
 use Maduser\Argon\Container\ArgonContainer;
 use Maduser\Argon\Container\Exceptions\ContainerException;
-use Maduser\Argon\Middleware\Factory\RequestHandlerFactory;
 use Maduser\Argon\Middleware\Contracts\MiddlewareLoaderInterface;
 use Maduser\Argon\Middleware\Contracts\MiddlewarePipelineCacheInterface;
 use Maduser\Argon\Middleware\Contracts\MiddlewareResolverInterface;
 use Maduser\Argon\Middleware\Contracts\PipelineManagerInterface;
 use Maduser\Argon\Middleware\Contracts\PipelineStoreInterface;
 use Maduser\Argon\Middleware\Contracts\RequestHandlerFactoryInterface;
+use Maduser\Argon\Middleware\Factory\RequestHandlerFactory;
 use Maduser\Argon\Middleware\Loader\TaggedMiddlewareLoader;
 use Maduser\Argon\Middleware\MiddlewarePipeline;
 use Maduser\Argon\Middleware\MiddlewarePipelineCache;
@@ -22,7 +22,7 @@ use Maduser\Argon\Middleware\Resolver\ContainerMiddlewareResolver;
 use Maduser\Argon\Middleware\Store\ContainerStore;
 use Psr\Http\Server\RequestHandlerInterface;
 
-final class RequestHandlerServiceProvider extends AbstractServiceProvider
+final class MiddlewarePipelineServiceProvider extends AbstractServiceProvider
 {
     private const DEFAULT_MIDDLEWARE_TAG = 'middleware.http';
 
@@ -50,9 +50,6 @@ final class RequestHandlerServiceProvider extends AbstractServiceProvider
 
         $container->set(MiddlewareResolverInterface::class, ContainerMiddlewareResolver::class);
 
-        /**
-         * Override the default middleware pipeline
-         */
         $container->set(RequestHandlerFactory::class);
 
         $container->set(RequestHandlerFactoryInterface::class, RequestHandlerFactory::class);